refactor: school -> universityName 이름 변경
