--- week04/kyulim/everytime/src/main/java/com/example/everytime/School/domain/repository/SchoolRepository.java
@@ -1,7 +1,0 @@
-package com.example.everytime.School.domain.repository;
-
-import com.example.everytime.School.domain.School;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SchoolRepository extends JpaRepository<School, Long> {
-}
