+++ week04/kyulim/everytime/src/main/java/com/example/everytime/board/domain/Board.java
@@ -1,7 +1,7 @@
 package com.example.everytime.board.domain;
 
-import com.example.everytime.Post.domain.Post;
-import com.example.everytime.School.domain.School;
+import com.example.everytime.post.domain.Post;
+import com.example.everytime.universityName.domain.UniversityName;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -33,7 +33,7 @@ public class Board {
     private List<Post> posts = new ArrayList<>();
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "school_id")
-    private School school;
+    @JoinColumn(name = "university_name_id")
+    private UniversityName universityName;
 
 }
