+++ week04/kyulim/everytime/src/main/java/com/example/everytime/universityName/domain/repository/UniversityNameRepository.java
@@ -1,0 +1,7 @@
+package com.example.everytime.universityName.domain.repository;
+
+import com.example.everytime.universityName.domain.UniversityName;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UniversityNameRepository extends JpaRepository<UniversityName, Long> {
+}
